updated with working destroy authorization

+++ app/controllers/likes_controller.rb
@@ -8,7 +8,11 @@ class LikesController < ApplicationController
   end
 
   def destroy
-    like = Like.where(user_id:params[:user_id], secret_id:params[:secret_id]).last.destroy
-    redirect_to secrets_index_path
+    if session[:user_id] == params[:user_id]
+      like = Like.where(user_id:params[:user_id], secret_id:params[:secret_id]).last.destroy
+      redirect_to secrets_index_path
+    else
+      redirect_to sessions_new_path
+    end
   end
 end
